#115 [move] : move NewsInfoModel to model module
