+++ core/model/src/main/java/com/teamwable/model/news/NewsInfoModel.kt
@@ -1,4 +1,4 @@
-package com.teamwable.news.model
+package com.teamwable.model.news
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
