+++ feature/news/src/main/java/com/teamwable/news/notice/NewsNoticeFragment.kt
@@ -5,9 +5,9 @@ import androidx.annotation.RequiresApi
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.navigation.fragment.findNavController
 import com.teamwable.designsystem.theme.WableTheme
+import com.teamwable.model.news.NewsInfoModel
 import com.teamwable.news.NewsFragmentDirections
 import com.teamwable.news.databinding.FragmentNewsNoticeBinding
-import com.teamwable.news.model.NewsInfoModel
 import com.teamwable.ui.base.BindingFragment
 import dagger.hilt.android.AndroidEntryPoint
