+++ feature/news/src/main/java/com/teamwable/news/notice/NewsNoticeRoute.kt
@@ -17,9 +17,9 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.teamwable.designsystem.theme.WableTheme
+import com.teamwable.model.news.NewsInfoModel
 import com.teamwable.news.NewsViewModel
 import com.teamwable.news.R
-import com.teamwable.news.model.NewsInfoModel
 
 @RequiresApi(Build.VERSION_CODES.O)
 @Composable
